Refactor event and add eventBus logic
